Adjusted site for react router.
