+++ src/layout/layout.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { Outlet } from "react-router-dom";
+
+export const Layout: React.FC = () => {
+	return (
+		<>
+			<main>
+				<Outlet />
+			</main>
+		</>
+	);
+};
